ecommerce_dashboard/requirements.txt
pandas>=1.3.0
matplotlib>=3.4.0
seaborn>=0.11.1
google-cloud-bigquery>=3.0.0
streamlit>=1.10.0
plotly>=5.0.0